requirements.txt

flask
requests
flask-socketio
eventlet
python-dotenv
langchain
openai
langchain-community
langchain-openai